requirements.txt
streamlit==1.32.0
openai>=1.56.0
pytesseract==0.3.10
Pillow==10.2.0
